requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.15
aiosignal==1.3.2
attrs==25.3.0
beautifulsoup4==4.13.3
bs4==0.0.2
certifi==2025.1.31
charset-normalizer==3.4.1
fake_user_agent==2.3.9
frozenlist==1.5.0
idna==3.10
lxml==5.3.1
multidict==6.3.0
propcache==0.3.1
python-dotenv==1.1.0
redis==5.2.1
requests==2.32.3
soupsieve==2.6
typing_extensions==4.13.0
urllib3==2.3.0
yarl==1.18.3
